Abstract unmarshal logic into helper function
