+++ src/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"log"
 
@@ -14,19 +13,6 @@ import (
 
 // isDebugLoggingEnabled toggles debug logging on if true, and false otherwise. Is read from the respective Lambda env var.
 var isDebugLoggingEnabled bool
-
-type RecordBody struct {
-	Method      string          `json:"method"`
-	Body        tgbotapi.Update `json:"body-json"`
-	QueryParams QueryParams     `json:"queryParams"`
-	PathParams  PathParms       `json:"pathParams"`
-}
-
-type QueryParams struct {
-}
-
-type PathParms struct {
-}
 
 func HandleRequest(ctx context.Context, event *events.SQSEvent) error {
 	if event == nil {
@@ -50,25 +36,11 @@ func HandleRequest(ctx context.Context, event *events.SQSEvent) error {
 	logInManualDebugMode("Number of records in event: %d", len(event.Records))
 	logInManualDebugMode("Event(s): %+v", event.Records)
 
-	for _, record := range event.Records {
-		var recordBodyUnmarshalDestination RecordBody
-
-		logInManualDebugMode("Event record output pre-unmarshal: %+v", recordBodyUnmarshalDestination)
-		logInManualDebugMode("Event record input pre-unmarshal: %s", record.Body)
-
-		// slice off last char in record.Body; record.Body is invalid json due to extra lagging " char
-		recordBodySourceSnipped := record.Body[:len(record.Body)-1]
-		logInManualDebugMode("Event record input pre-unmarshal snipped: %s", record.Body)
-
-		if err := json.Unmarshal([]byte(recordBodySourceSnipped), &recordBodyUnmarshalDestination); err != nil {
-			log.Printf("error when unmarshaling Telegram Update object: %v", err)
-			continue
+	for _, sqsMessage := range event.Records {
+		update, err := getTelegramUpdateFromSQSMessage(sqsMessage)
+		if err != nil {
+			log.Printf("error when unmarshaling SQS message: %v", err)
 		}
-
-		logInManualDebugMode("Event record output post-unmarshal: %+v", recordBodyUnmarshalDestination)
-
-		update := recordBodyUnmarshalDestination.Body
-		logInManualDebugMode("Update: %+v", update)
 
 		// if _, err := bot.Send(tgbotapi.NewMessage(update.Message.Chat.ID, update.Message.Text)); err != nil {
 		chattable := tgbotapi.NewMessage(update.Message.Chat.ID, update.Message.Text)
